Fix the miscount of blames in --all
The problem comes from the fact that even when listing all files
appearing in all commits, it is still not easy to find all the
interesting files in the repository. For example, the file renames
still seem to be lost from time to time.

So this reverts to the previous behavior, but there is still a
need for finding these files in the information given by "git log".